Allow separate match / count delays for file rendering and change --grep-count to output matches immediately
As this is a frequent use case: I'm interested in accumulated counts at the end, but need any encountered match. The previous implementation could only report the final match at the end, or have all intermediate count increments reported as well.
By splitting the delays, we can handle this. delayedRender() is generalized into renderIt(), and the separate parts of delayedFileRender() are duplicated in postRenderLine() to handle differing delays.