deleteStackTraceLocations: Avoid referring to last capture group
Because of the previously established match, we don't actually need to capture and re-insert the last stacktrace line. The \n at the end of the pattern also assures that we're matching (and then discarding) the middle stacktrace line, and that another line follows (which previously was captured, now is outside of the match). This lifts the dependency that the DELETESTACKTRACELOCATIONS_OUR_PATTERN has exactly one capture group.