Documentation: Tweak --summarize functionality
Combining --tally and --summarize would be difficult to implement, as both would need the record continuation algorithm - with different filtering (and one printing records, and the latter storing for a final summary). Also, the separation of both outputs is awkward to handle, and would make parsing of that output cumbersome.
I think it's better to limit the application's ability to one task; if both are needed, it would have to be invoked multiple times on the same source files (or using tee). That would have been necessary for multiple summaries or tallies, anyway.