Adapt --to overlay --quiet formatting defaults
As the input text is not continually output, there's no scrolling, and we do not need to move to the terminal's first line; just repeatedly overwrite the current line.
Because the displayed matches / counts may get shorter, clear the text after the overlay until the end of the line.
As the information does not go on top of the previous input, we do not need the colorful highlighting here.