rot45: ENH: Suppress empty lines at the beginning and end
When the input hasn't a quadratic form, the rotation may lead to empty lines; this is ugly.
To suppress those at the beginning, we just need to skip output until the first non-empty line is encountered.
For the end, we need to stop direct output once an empty line is encountered (this is not the final end; more non-empty lines could follow!) and just accumulate in an array, which we then remove trailing empty lines from and output the remainder.