highlightRelativeAge: BUG: "6 days ago" is not removed by --delete-date
In the original git-dashboardagecoloring script, an age of a few days just fell through without coloring. Now that the script also (optionally) does the deletion, we need to match any date, even if that doesn't result in highlighting.
Add a pattern match branch for few days, skip other matches if one matches (sed's "t" command) to avoid multiple matches for few / many days (as sed does not allow lookbehind). Use the reset color for the new branch to be a no-op for highlighting.