FIX: Removal of individual sequences mangle 256-color sequence
For example, removal of "blink" (5) also mangles the 256-color sequence 38;5;11 into 38;11.
Sequential sequences need to be treated as one big sequence. We can achieve this by first turning the inner ; delimiters into something else (for the input as well as for any COLOR arguments), and at the end undo this.