FIX: Removal of individual sequences mangle RGB-color sequence
For example, removal of "background red" (41) also mangles the RGB sequence 38;2;41;33;41 into 38;2;33.
Sequential sequences need to be treated as one big sequence. We can achieve this by first turning the inner ; delimiters into something else (for the input as well as for any COLOR arguments), and at the end undo this.