FIX: Mask single (attribute and basic color) sequences only in replacements
To avoid endless loops when the same sequence is matched - this can easily happen with the color classes; e.g. "uncolor --backgrounds to red": the replacement (red) is also matched by --backgrounds, for sed an (equal) replacement happened, and it will continue to eternity. We could use the hold space to implement real change tracking, but it's easier to just mask the single-sequence replacement values with ^G^G, too (red: 41 -> 41^G^G), as the masking is already in place, and we just need to minimally extend it.