Also determine the sequence type when ATTRIBUTE|COLOR is passed
Without a prepended foreground / background, we have to check on our own, unfortunately, as "colored" only tells us the sequence, but not what type it is. We already duplicate the sequence pattern for attributes (in the --attributes branch), but doing a lookup through pattern matching is even more complex than duplicating the few (and hopefully static) attribute names here.