BUG: Broken for multi-line
There are several problems with the existing implementation; I apparently only tested this fleetingly. numfmt needs whitespace around the number, and the appending of the magic \d001 value already breaks that.
The separate tr pipeline steps can be done with sed.
Better use the bell ^G / \d007 instead; this is what I've decided on elsewhere, too.
Rework the algorithm.