Implement combination of --skip-empty and --join-by
With DWIM semantics (as far as I can see it right now): If the joiner TEXT does not contain a newline, interpret the empty line as a stopper just like the end of input, and don't include a joiner there. The empty line itself will be swallowed (empty-joined). If the joiner TEXT is a newline, this is a no-op. If the joiner TEXT contains a newline and other text, that would ideally be completely suppressed around an empty line, too. Unfortunately, any parts of the joiner up to the last newline have already been output when the empty line is encountered, so we can only remove what's left, after the last newline. I could not find a way around this, as the complete joiner cannot be put onto the hold space; any merge with pattern space will introduce another newline, and as the pattern space might have gotten newlines via prefix or suffix, there's no way to pick that particular newline. Likewise, the last line can be easily recognized via the $ address, but to recognize a following empty line we'd have to read it via N, but would have to put it back in order to move the remainder of the current line + joiner TEXT to the hold space.