Better handling of --skip-empty and --join-by
The last line can be easily recognized via the $ address; to recognize a following empty line we have to read it via N and try a match with $skipEmptyPattern. Continuing, we need to keep the previous line (i.e. any left-over from the surrounded original line and the join TEXT) and the last line, push the first to the hold space and keep the last in the pattern space, then restart.
Compared to the prior implementation, this will also keep the empty line intact when the join TEXT does not contain a newline. This is more consistent, but I find this less helpful, but empty lines can easily filtered out by the client, anyway.