FIX: Correctly handle joiner TEXT that contains a mix of newlines and other characters
The prefix parsing so far assumed that only the original line is in the pattern space; however, now there can be previous line(s) from joiner TEXT as well. Change ^ anchor to \(^\|\n\) and turn the [[:space:]] (which includes newline characters) into [ \t\v\f\r] so that we only ever capture the last line in pattern space. This avoids that the prefix includes text from the joiner.