extractMatches: TESTSFIX: Use $# -gt 1 to check fixture invocation
The check for $RUNS doesn't work (any longer? found this after updating to Bats 1.6.0). The fixture now is invoked twice; $RUNS is only undefined the first time.
Bats apparently passes the fixture filespec when invoking; as the test use pass multiple arguments, we can use the argument number to differentiate.