extractMatchesToPanel: Adapt: Go back to using WINDOWID instead of my own TERMID alternative, as there are (now) different meanings
For updating the terminal title, we need WINDOWID as that represents something that has a GUI window and therefore a title. A tmux session may have multiple windows consisting of potentially multiple panes, and each such shell would have its own TERMID, but there's just a single window for all of them.