Reimplementation: Obtain less precision by rounding smallest units until it fits
Now that convert() directly switches from HH:MM:SS to individual units when component(s) are 0, there's no need for the explicit conversion, and we can retry the conversion with an integer that has been rounded to the next unit.
The rounding has the big advantage that instead of truncating smaller units, we obtain the "nearest" value, so "1d 22:40:59" does not get truncated to "1d 22h 40m" → "1d 22h" → "1d", but instead to "1d 22h 41m" → "1d 23m" → "2d".