ENH: Implement skipping of matches via -s|--skip-regexp
Due to the dynamic re-anchoring of ^, it's useful to skip previous parts of the line in order to emulate a lookbehind assertion; for a single match, regexp grouping and --replacement can be used, but that doesn't work with multiple overlaying regexp matches.
Instead, an approach similar to Vim's :syn-nextgroup can be used: Match with a dummy pattern for the lookahead, and then (indiscriminately any) real match can be anchored with ^.
For the no-op match, a combination of -e $^ -x LOOKAHEAD-PATTERN can be used, but this is hacky.
Rather, offer a built-in no-op match (whose implementation is trivial).