ENH: Allow reporting percentages against canned fixed values passed via --percentage-total
This is useful when the total sum cannot be determined from the tallied matches. For example, when matching commit message keywords, we want the percentages for each keyword measured not against all commits with keywords, but against _all_ commits. With the primitive regular expressions, it's difficult to impossible to somehow match all other commit messages. (And then these non-matches somehow have to be excluded from the reporting.)