sampleLines: Documentation: Add example where output is piped through
One use case it just consuming the (abundant) output and condensing it to regular status updates.
But another use is inside a pipeline, sampling some of the data and directly printing it to the terminal to give an update of what's happening.
This is similar to pv (pipe viewer) usage, so mention that related command as well.

System-info: Armbian 24.11.1 noble, aarch64
Platform-info: bash 5.2.21(1)-release, Bats 1.10.0