ENH: DWIM: Allow line filtering of --match and --skip via --only-partials|-P
Lines with partial matches may often be the only ones we're interested in. So far, non-matching lines would have to be explicitly excluded via a duplicate --grep PATTERN, which isn't nice.
Add a flag that makes any line that doesn't have a partial match ineligible for recording and have it directly printed, as if excluded by --vgrep.

This new feature make me realize the problems with whitespace handling and the intricacies of the existing filtering.

System-info: Ubuntu 20.04.6 LTS, x86_64
Platform-info: bash 5.0.17(1)-release, Bats 1.10.0